requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-community
pypdf
chromadb==0.4.22
python-docx
pillow
python-dotenv
gtts
docx2txt
numpy==1.26.4
azure-cognitiveservices-speech
st-audiorecorder
SpeechRecognition